fix(apps/web): use node:http HMAC gateway instead of ws bridge
Promote scripts/collab-gateway.mjs as the public Upgrade edge for split
web/collab hosts, drop @vercel/functions and ws, and keep proxy.ts as a
signed rewrite fallback only.
