+++ apps/web/.env.example
@@ -5,6 +5,12 @@ COLLAB_BACKEND_ORIGIN="http://localhost:3002"
 COLLAB_GATEWAY_HMAC_KEY_ID="local-dev"
 COLLAB_GATEWAY_HMAC_SECRET="AAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHh8"
 
+# scripts/collab-gateway.mjs (public edge in front of Next)
+# NEXT_ORIGIN="http://127.0.0.1:3001"
+# COLLAB_GATEWAY_PORT="3000"
+# COLLAB_GATEWAY_PUBLIC_ORIGIN="http://127.0.0.1:3000"
+# COLLAB_GATEWAY_BIND="127.0.0.1"
+
 # Only for an isolated E2E project. The seed route returns 404 unless every
 # guard is present and E2E_SUPABASE_PROJECT_REF matches the URL exactly.
 E2E_ALLOW_REMOTE_SEED="false"
