+++ apps/web/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "next dev",
     "dev:e2e": "next dev",
+    "gateway": "node ./scripts/collab-gateway.mjs",
     "e2e:gateway": "node ./scripts/e2e-collab-gateway.mjs",
     "build": "next build",
     "start": "next start",
@@ -35,7 +36,6 @@
     "@softmaple/ui": "workspace:*",
     "@supabase/ssr": "^0.12.4",
     "@supabase/supabase-js": "^2.112.2",
-    "@vercel/functions": "^3.9.1",
     "dayjs": "^1.11.21",
     "dompurify": "^3.4.12",
     "lexical": "^0.44.0",
@@ -48,7 +48,6 @@
     "react-markdown": "10.1.0",
     "remark-gfm": "4.0.1",
     "sharp": "0.35.3",
-    "ws": "^8.21.3",
     "zod": "^4.3.5"
   },
   "devDependencies": {
@@ -59,7 +58,6 @@
     "@types/node": "^24.10.4",
     "@types/react": "19.2.18",
     "@types/react-dom": "19.2.4",
-    "@types/ws": "^8.18.1",
     "eslint": "^10.8.0",
     "jsdom": "^30.0.1",
     "typescript": "^5.9.3",
