+++ turbo.json
@@ -1,18 +1,22 @@
 {
   "$schema": "https://v2-9-12.turborepo.dev/schema.json",
-  "globalEnv": ["CI", "VERCEL"],
+  "globalEnv": ["CI"],
   "tasks": {
     "build": {
       "dependsOn": ["^build"],
       "env": [
         "COLLAB_BACKEND_ORIGIN",
+        "COLLAB_GATEWAY_BIND",
         "COLLAB_GATEWAY_HMAC_KEY_ID",
         "COLLAB_GATEWAY_HMAC_KEYS",
         "COLLAB_GATEWAY_HMAC_SECRET",
+        "COLLAB_GATEWAY_PORT",
+        "COLLAB_GATEWAY_PUBLIC_ORIGIN",
         "DATABASE_URL",
         "E2E_ALLOW_REMOTE_SEED",
         "E2E_SEED_SECRET",
         "E2E_SUPABASE_PROJECT_REF",
+        "NEXT_ORIGIN",
         "NEXT_PUBLIC_APP_URL",
         "NEXT_PUBLIC_SUPABASE_ANON_KEY",
         "NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY",
@@ -45,6 +49,9 @@
     "test": {
       "dependsOn": ["^build"],
       "env": [
+        "COLLAB_GATEWAY_BIND",
+        "COLLAB_GATEWAY_PORT",
+        "COLLAB_GATEWAY_PUBLIC_ORIGIN",
         "E2E_BASE_URL",
         "E2E_COLLAB_ORIGIN",
         "E2E_COLLAB_PORT",
@@ -54,7 +61,8 @@
         "E2E_SEED_SECRET",
         "E2E_WEB_PORT",
         "EG_WALKER_PROPERTY_RUNS",
-        "EG_WALKER_REFERENCE_ROOT"
+        "EG_WALKER_REFERENCE_ROOT",
+        "NEXT_ORIGIN"
       ]
     },
     "dev": {
